requirements.txt
# Lotus Trader - Organic Intelligence Trading System
# Complete dependency list for all teams and components

# =============================================================================
# CORE DEPENDENCIES
# =============================================================================

# Database
psycopg2-binary==2.9.10
asyncpg==0.29.0
sqlalchemy==2.0.25

# WebSocket and HTTP
websockets==15.0.1
aiohttp==3.9.1
httpx==0.26.0
requests==2.31.0

# Async support
asyncio==4.0.0

# Environment management
python-dotenv==1.1.1

# =============================================================================
# DATA PROCESSING & ANALYSIS
# =============================================================================

# Core data processing
pandas==2.2.0
numpy==1.26.4
scipy==1.11.4

# Technical analysis
ta-lib==0.4.28
yfinance==0.2.28

# Machine learning (for pattern detection)
scikit-learn==1.3.2
scikit-optimize==0.9.0

# =============================================================================
# LLM INTEGRATION
# =============================================================================

# OpenAI and Anthropic
openai==1.12.0
anthropic==0.8.1

# Vector embeddings and similarity
sentence-transformers==2.2.2
chromadb==0.4.18

# =============================================================================
# TRADING & MARKET DATA
# =============================================================================

# Cryptocurrency exchanges
ccxt==4.1.77

# Market data processing
pytz==2023.3
python-dateutil==2.8.2

# Trading APIs and blockchain
solana==0.30.2
web3==6.11.3
eth-account==0.9.0

# Web scraping for social media
playwright==1.48.0
telethon==1.36.0

# Supabase for database
supabase==2.18.1

# System monitoring
psutil==5.9.8

# =============================================================================
# CONFIGURATION & LOGGING
# =============================================================================

# Configuration
PyYAML==6.0.2
pydantic==2.5.2
pydantic-settings==2.1.0

# Logging
structlog==24.1.0
loguru==0.7.2

# =============================================================================
# DEVELOPMENT & TESTING
# =============================================================================

# Testing
pytest==8.2.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Type checking
types-requests==2.31.0.10
types-PyYAML==6.0.12.12

# =============================================================================
# MONITORING & OBSERVABILITY
# =============================================================================

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.39.2

# Performance profiling
memory-profiler==0.61.0
line-profiler==4.1.2

# =============================================================================
# SECURITY
# =============================================================================

# Security
cryptography==41.0.8
pyjwt==2.8.0

# =============================================================================
# UTILITIES
# =============================================================================

# Utilities
click==8.1.7
rich==13.7.0
tqdm==4.66.1
python-multipart==0.0.6

# Caching
redis==5.0.1
aioredis==2.0.1

# =============================================================================
# OPTIONAL DEPENDENCIES (for advanced features)
# =============================================================================

# Jupyter notebooks (for analysis)
jupyter==1.0.0
ipykernel==6.29.0
matplotlib==3.8.2
seaborn==0.13.0

# Advanced ML (optional)
torch==2.1.2
transformers==4.36.2

# Note: Some packages may require system dependencies
# Install with: pip install -r requirements.txt
